Name changes + moving of code
